Created concurrency test and refactored all tests

+++ userspace/test_util.h
@@ -1,5 +1,5 @@
 #include <stdio.h>
-#include "hashtable.h"
+#include "userspace/hashtable.h"
 
 void fail(void);
 
@@ -7,6 +7,6 @@ void putval(int key, int val);
 
 uint8_t * getval(int key);
 
-void removval(uint8_t key);
+void removval(int key);
 
 int compareStrings(uint8_t * v1 , uint8_t *v2, ssize_t len);
